6649 turntable improvements (#2)
Cleaned up existing code which was very hard to follow and understand.
Added the ability to render the turntable with Unreal Movie render queue, if available. The script checks if a render manifest was created to see if it can use the Movie Render queue or not.
Make a copy of the Unreal Project in a temp folder before editing it for the turntable setup.
Pre-loaded needed plugins in the 2.6 Unreal template Project.